# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "AreaOut": {
        "properties": {
          "geo_area_id": {
            "title": "Geo Area Id",
            "type": "integer"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "teryt_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teryt Code"
          },
          "weather_polling_active": {
            "title": "Weather Polling Active",
            "type": "boolean"
          }
        },
        "required": [
          "geo_area_id",
          "slug",
          "name",
          "teryt_code",
          "latitude",
          "longitude",
          "weather_polling_active"
        ],
        "title": "AreaOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaceAreaResponse": {
        "properties": {
          "area": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AreaOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "place": {
            "$ref": "#/components/schemas/PlaceOut"
          },
          "polling": {
            "enum": [
              "active",
              "inactive",
              "capacity_reached",
              "budget_exhausted"
            ],
            "title": "Polling",
            "type": "string"
          }
        },
        "required": [
          "place",
          "area",
          "polling",
          "attribution"
        ],
        "title": "PlaceAreaResponse",
        "type": "object"
      },
      "PlaceOut": {
        "properties": {
          "admin1_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin1 Code"
          },
          "admin2_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin2 Code"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "place_id": {
            "title": "Place Id",
            "type": "integer"
          },
          "population": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population"
          }
        },
        "required": [
          "place_id",
          "name",
          "kind",
          "admin1_code",
          "admin2_code",
          "latitude",
          "longitude",
          "population"
        ],
        "title": "PlaceOut",
        "type": "object"
      },
      "PlacesResponse": {
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "places": {
            "items": {
              "$ref": "#/components/schemas/PlaceOut"
            },
            "title": "Places",
            "type": "array"
          }
        },
        "required": [
          "places",
          "attribution"
        ],
        "title": "PlacesResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Za Oknem API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/places": {
      "get": {
        "description": "Search the places registry by name prefix, diacritics-insensitive (ADR-029): any\nlocality in Poland, not only cities with a station. Exact match first, then by\npopulation. Reads only our own DB (rule #14); the app itself never logs `q` (only the path).",
        "operationId": "list_places_api_v1_places_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Places"
      }
    },
    "/api/v1/places/{place_id}": {
      "get": {
        "description": "A place and the state of its area. Read-only: never creates or refreshes the area\n(the id is public and enumerable, so a read must not keep polling alive, ADR-026).",
        "operationId": "get_place_api_v1_places__place_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "title": "Place Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceAreaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Place"
      }
    },
    "/api/v1/places/{place_id}/activate": {
      "post": {
        "description": "Choose a place: create its area on first use and (re)start collecting data for it\n(ADR-029). Idempotent; refreshes the idle-expiry clock - the client calls it when the\nuser selects the place and when the app opens with it selected. Data arrives with the\nscheduler (first fetch within minutes, not the 3 h cycle), never on this request\n(rule #14). No user data is sent or stored: the area is a public resource (rule #11).\nSwitching polling on is rate-limited per IP; refreshing an already active area is not.",
        "operationId": "activate_api_v1_places__place_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "title": "Place Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceAreaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate"
      }
    },
...
  }
}